# top-level/output: add enablePrintInit option